+++ services/repository/src/main/java/org/sagebionetworks/repo/web/service/UserProfileServiceImpl.java
@@ -234,7 +234,7 @@ public class UserProfileServiceImpl implements UserProfileService {
 	
 	@Override
 	public PaginatedResults<ProjectHeader> getProjects(Long userId, Long otherUserId, Long teamId, ProjectListType type,
-			ProjectListSortColumn sortColumn, SortDirection sortDirection, Integer limit, Integer offset) throws DatastoreException,
+			ProjectListSortColumn sortColumn, SortDirection sortDirection, Long limit, Long offset) throws DatastoreException,
 			InvalidModelException, NotFoundException {
 		UserInfo userInfo = userManager.getUserInfo(userId);
 		UserInfo userToGetInfoFor = userInfo;
